整合微信公众平台开发工具

+++ opendev-server-api-impl/opendev-user-server-api-impl/src/main/java/com/opendev/mapper/UserMapper.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 public interface UserMapper extends BaseMapper<User> {
 
-    User selectById(@Param("userId") Integer userId);
+    User selectById(@Param("userId") Long userId);
 
     User selectByUsername(@Param("username") String username);
